+++ R/repair-names.R
@@ -8,11 +8,11 @@
 #'   * `unique` names are `minimal`, have no duplicates, and are never empty
 #'     (literally, no `""`s).
 #'     - All columns can be accessed by name via `df[["name"]]`.
-#'   * `syntactic` names are `unique` and syntactic (see Details for more).
+#'   * `universal` names are `unique` and syntactic (see Details for more).
 #'     - Names work everywhere, without quoting: `df$name` and
 #'     `lm(name1 ~ name2, data = df)` and `dplyr::select(df, name)` all work.
 #'
-#' `syntactic` implies `unique`, `unique` implies `minimal`. These levels are
+#' `universal` implies `unique`, `unique` implies `minimal`. These levels are
 #' nested.
 #'
 #' The `.name_repair` argument of [tibble()] and [as_tibble()] refers to these
@@ -61,9 +61,9 @@
 #' names `unique`, i.e. reconstructing the suffixes. If this interacts poorly
 #' with your names, you should take control of name repair.
 #'
-#' @section `syntactic` names:
+#' @section `universal` names:
 #'
-#' `syntactic` names are `unique` and syntactic, meaning they:
+#' `universal` names are `unique` and syntactic, meaning they:
 #'   * Are never empty (inherited from `unique`).
 #'   * Have no duplicates (inherited from `unique`).
 #'   * Consist of letters, numbers, and the dot `.` or underscore `_`
@@ -72,7 +72,7 @@
 #'   * Are not a [reserved] word, e.g., `if` or `function` or `TRUE`.
 #'   * Are not `...`. Do not have the form `..i`, where `i` is a number.
 #'
-#' If a data frame has `syntactic` names, variable names can be used "as is" in
+#' If a data frame has `universal` names, variable names can be used "as is" in
 #' code. They work well with nonstandard evaluation, e.g., `df$name` works.
 #'
 #' Tibble has a different method of making names syntactic than
@@ -82,10 +82,10 @@
 #' Examples:
 #' ```
 #'  Original names:     ""    "x"    NA     "x"
-#' syntactic names: "...1" "x..2" "...3" "x..4"
+#' universal names: "...1" "x..2" "...3" "x..4"
 #'
 #'   Original names: "(y)"  "_z"  ".2fa"  "FALSE"  "..."  "..3"
-#'  syntactic names: ".y." "._z" "..2fa" ".FALSE" "...." "...6"
+#'  universal names: ".y." "._z" "..2fa" ".FALSE" "...." "...6"
 #' ```
 #'
 #' @param x A vector.
@@ -113,7 +113,7 @@
 #' df$`a 1`
 #'
 #' ## syntactic names are easier to work with, though, and you can request them
-#' df <- tibble(`a 1` = 1, `a 2` = 2, .name_repair = "syntactic")
+#' df <- tibble(`a 1` = 1, `a 2` = 2, .name_repair = "universal")
 #' df$a.1
 #'
 #' ## you can specify your own name repair function
@@ -132,14 +132,14 @@
 NULL
 
 set_repaired_names <- function(x,
-                               .name_repair = c("check_unique", "unique", "syntactic", "minimal")) {
+                               .name_repair = c("check_unique", "unique", "universal", "minimal")) {
   x <- set_minimal_names(x)
   names(x) <- repaired_names(names(x), .name_repair = .name_repair)
   x
 }
 
 repaired_names <- function(name,
-                           .name_repair = c("check_unique", "unique", "syntactic", "minimal")) {
+                           .name_repair = c("check_unique", "unique", "universal", "minimal")) {
 
   if (is_formula(.name_repair, lhs = FALSE)) {
     .name_repair <- as_function(.name_repair)
@@ -157,14 +157,14 @@ repaired_names <- function(name,
       minimal       =     ,
       check_unique  = NULL,
       unique        = unique_names,
-      syntactic     = syntactic_names,
+      universal     = universal_names,
       abort(error_name_repair_arg())
     )
   }
   new_name <- if (is_function(repair_fun)) repair_fun(name) else name
 
   if (is.character(.name_repair) &&
-    .name_repair %in% c("check_unique", "unique", "syntactic")) {
+    .name_repair %in% c("check_unique", "unique", "universal")) {
     check_unique(new_name)
   } else {
     check_minimal(new_name)
@@ -212,13 +212,13 @@ set_unique_names <- function(x, quiet = FALSE) {
   set_names(x, new_names)
 }
 
-syntactic_names <- function(name, quiet = FALSE) {
+universal_names <- function(name, quiet = FALSE) {
   unique_names(name, quiet = quiet, transform = make_syntactic)
 }
 
-set_syntactic_names <- function(x, quiet = FALSE) {
+set_universal_names <- function(x, quiet = FALSE) {
   x <- set_minimal_names(x)
-  new_names <- syntactic_names(names(x), quiet = quiet)
+  new_names <- universal_names(names(x), quiet = quiet)
   set_names(x, new_names)
 }
 
@@ -356,10 +356,10 @@ describe_repair <- function(orig_name, name) {
 #' `as_tibble()` did not do this.
 #'
 #' From tibble v1.5.0, the `.name_repair` argument gives direct access to three
-#' specific levels of name repair: `minimal`, `unique`, and `syntactic`. We
+#' specific levels of name repair: `minimal`, `unique`, and `universal`. We
 #' recommend that new code use this instead of `tidy_names()`,
 #' `set_tidy_names()`, or `repair_names()`. After a period of use, the repair
-#' stategies behind `minimal`, `unique`, and `syntactic` are likely to be
+#' stategies behind `minimal`, `unique`, and `universal` are likely to be
 #' exposed in standalone functions and this could affect the behaviour of
 #' `tidy_names()`. `repair_names()` should be considered deprecated.
 #'
@@ -369,7 +369,7 @@ describe_repair <- function(orig_name, name) {
 #' df <- tibble(...)
 #' set_tidy_names(df, syntactic = FALSE)
 #'
-#' tibble(..., `.name_repair = "syntactic"`)
+#' tibble(..., `.name_repair = "universal"`)
 #' ## is preferred to
 #' df <- tibble(...)
 #' set_tidy_names(df, syntactic = TRUE)
@@ -377,12 +377,12 @@ describe_repair <- function(orig_name, name) {
 #'
 #' @param syntactic Should names be made syntactically valid? If `FALSE`, uses
 #'   same logic as `.name_repair = "unique"`. If `TRUE`, uses same logic as
-#'   `.name_repair = "syntactic"`.
+#'   `.name_repair = "universal"`.
 #' @export
 #' @rdname name-repair
 tidy_names <- function(name, syntactic = FALSE, quiet = FALSE) {
   if (syntactic) {
-    syntactic_names(name, quiet = quiet)
+    universal_names(name, quiet = quiet)
   } else {
     unique_names(name, quiet = quiet)
   }